% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reductions.R
\name{aaReduction}
\alias{aaReduction}
\title{Reduce TCR data based on amino acid characteristics}
\usage{
aaReduction(
  sc,
  chains = "TRB",
  AA.properties = "AF",
  AA.method = "auto",
  n.trim = 0,
  c.trim = 0
)
}
\arguments{
\item{sc}{Single Cell Object in Seurat or SingleCell Experiment format}

\item{chains}{TRA or TRB (no both option)}

\item{AA.properties}{Amino acid properties to use for distance calculation: 
"AF" = Atchley factors, "KF" = Kidera factors, "both" = AF and KF, or NULL to not 
include edit distance layer in multiplex network}

\item{AA.method}{method for deriving pairwise distances from amino acid properties, 
either "mean" or the average amino acid characteristics for given cdr3 sequence or "auto"
for using an autoencoder model. The autoencoder selection  will not allow for trimming of 
the cdr3 sequences.}

\item{n.trim}{length from N-terminus to trim on the CDR3 aa sequence}

\item{c.trim}{length from C-terminus to trim on the CDR3 aa sequence}
}
\value{
30-dimensional autenocoder bottleneck layer or mean values for each cell
}
\description{
Use this to get the vector returned from the TCR autoencoder 
or mean values based on indicated amino acid property before
the network generation.
}
\examples{
AA.dim  <- aaReduction(trex_example, 
                        chains = "TRB", 
                        AA.properties = "AF",
                        AA.method = "auto",
                        n.trim = 0,
                        c.trim = 0)

}
