% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.data.R
\name{agg.data}
\alias{agg.data}
\title{Aggregation of time-series data}
\usage{
(input,
 time.agg = 60*24,
 start = "2012-07-28 00:00",
 end = "2012-10-29 00:00",
 FUN = "sum",
 unit = 60,
 na.rm = TRUE,
 df = FALSE)
}
\arguments{
\item{input}{An \code[is.trex]-compliant time series from \code{cal.sfd} outputs
(e.g., \code(X$sfd.mw$sfd)).}

\item{time.agg}{Numeric, the aggregation time in minutes (default = 60).}

\item{start}{Character string, the start time for the series. Format has
to be provided in "UTC" (e.g. "2012-05-28 00:00" or Year-Month-Day Hour:Minute).
Starting time should not be earlier than the start of the series.
If not provided the entire series is considered.}

\item{end}{character string, the end time for the series.
Format has to be provided in "UTC" (e.g. "2012-06-28 00:50" or Year-Month-Day Hour:Minute).
Starting time should be earlier than the end time and the end time should not exceed the
length of the series.  If not provided the entire series is considered.}

\item{FUN}{Quoted function name to compute the summary statistics which can be
applied to all data subsets (see aggregate; including “sum", "mean",
 "median", "sd", "se", "min", "max").}

\item{unit}{Numeric, the minutes in which a velocity unity is provided
(e.g., cm3 cm-2 h-1 = 60) for summation (\code{FUN = “sum”}; default = 60).}

\item{na.rm}{Logical; iff \code{TRUE} missing values are removed (default = \code{TRUE}).}

\item{df}{Logical; If \code{TRUE}, output is provided in a \code{data.frame} format
with a timestamp and a value column. If \code{FALSE}, output
is provided as a zoo vector object (default = \code{FALSE}).}
}
\value{
A zoo object or data.frame in the appropriate format for other functionalities.
}
\description{
Aggregation of time-series data and start/end time selection.
This function provides the option to select the temporal step size
for aggregation of a single time series origination from an \code{is.trex}-compliant object.
Additionally, the user can define the start and end time of the series and select
the mean of aggregating the data.
}
