% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{alignment_wrapper}
\alias{alignment_wrapper}
\title{Count the 'good' and 'total' informative cases between input peak set with reference database.}
\usage{
alignment_wrapper(input_vec, bin_width, option)
}
\arguments{
\item{input_vec}{A input vector contains index of peaks transformed by applying import_peaks.}

\item{bin_width}{width of bin, should be in 100/500/1000.}
}
\value{
a data frame has three columns, TR labels, number of 'good' windows, number of 'total' informative cases.
}
\description{
Count the 'good' and 'total' informative cases between input peak set with reference database.
}
