% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allometry.data}
\alias{allometry.data}
\title{Sapwood area and depth, and bark allometric measurements}
\format{
Provides a data.frame with 651  rows and 8 columns.
\describe{
  \item{Plot}{Measurement plot (\code{character})}
  \item{treeID}{Unique tree ID (\code{character})}
  \item{Species}{Assessed species (\code{character})}
  \item{DBH_cm}{Diameter at breast height (1.3 \eqn{m}) in \eqn{cm} (\code{numeric})}
  \item{Bark_cm}{Bark thickness in \eqn{cm} (\code{numeric})}
  \item{Sap_cm}{Depth of sapwood in \eqn{cm} (\code{numeric})}
  \item{Saparea_cm2}{Sapwood area in \eqn{cm^2} (\code{numeric})}
  \item{Site}{Measurement site \code{character})}
}
}
\usage{
allometry.data
}
\description{
Returns raw allometric data obtained for 14 spepcies from several North-American sites.
 These data are used to develop allometric equations for the parameters sapwood area and depth, as well as bark thickness.
 A detailed description on collection method and sites is provided in Pappas \emph{et al.} (2021; in prep.).
 The parameters were measured along an increment core extracted at breast height (DBH, 1.3 \eqn{m}) with a Haglof borer.
 Bark thickness (\code{Bark_cm}) was measured directly, while sapwood depth (\code{Sap_cm}) was estimated through discoloration and translucence.
 Sapwood area (\code{Saparea_cm2} was then calculated using:
 \deqn{Saparea_cm = \pi (DBH/2 - Bark_cm)^2  - \pi (DBH/2 - Bark_cm - Sap_cm)}
}
\details{
Currently included species are:

\itemize{
     \item {Abies balsamea}
     \item {Acer rubrum}
     \item {Acer saccharum}
     \item {Betula alleghaniensis}
     \item {Betula papyrifera}
     \item {Fagus grandifolia}
     \item {Fraxinus nigra}
     \item {Picea glauca}
     \item {Picea mariana}
     \item {Picea rubens}
     \item {Pinus banksiana}
     \item {Populus tremuloides}
     \item {Quercus rubra}
     \item {Thuja occidentalis}

}
}
\keyword{datasets}
