% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allometry.models}
\alias{allometry.models}
\title{Sapwood area and depth, and bark allometric models}
\format{
Provides a list (one item for each allometric parameter)
with model data.frames/tibbles nested by species.
Relevant parameters can be extracted either by unnesting (\code{tidyr::unnest()}), or by
selecting a single nested item, e.g., \code{allometric.models$Saparea_cm2$fit[[1]]}.

Each data.frame/tibble has four columns

\describe{
  \item{Species}{Assessed species (\code{character})}
  \item{data}{data used to fit model;
  single-species subset of \code{\link{allometry.data}} (\code{list column})}
  \item{fit}{nls or lm model fit (\code{list column})}
  \item{model_parameters}{Convenient summary of models as produced by \code{broom::tidy()} (\code{list column})}
}
}
\usage{
allometry.models
}
\description{
Returns allometric models for 14 spepcies from several North-American sites.
 The models relate independent (i.e. outcomes) parameters sapwood area and depth, as well as bark thickness
 to diameter at breast height (DBH, 1.3 \eqn{m}).
 A detailed description on model development and assessment is provided in Pappas \emph{et al.} (2021; in prep.).
 The data used for these models is found in \code{\link{allometry.data}}.
}
\details{
OLS models (y ~ x) are used for relationships between DBH (dependent) and sapwood / bark thickness;
A non-linear LS model (y ~ a * x ^ b ) is used for the relationship between DBH (dependent) and sapwood area.

\itemize{
     \item {Abies balsamea}
     \item {Acer rubrum}
     \item {Acer saccharum}
     \item {Betula alleghaniensis}
     \item {Betula papyrifera}
     \item {Fagus grandifolia}
     \item {Fraxinus nigra}
     \item {Picea glauca}
     \item {Picea mariana}
     \item {Picea rubens}
     \item {Pinus banksiana}
     \item {Populus tremuloides}
     \item {Quercus rubra}
     \item {Thuja occidentalis}

}
}
\keyword{datasets}
