% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDB.R
\name{annotateDB}
\alias{annotateDB}
\title{Use reference database to annotate cdr3 sequences}
\usage{
annotateDB(sc, chains = "TRB", edit.distance = 0)
}
\arguments{
\item{sc}{Single Cell Object in Seurat or SingleCell Experiment format}

\item{chains}{TRA or TRB}

\item{edit.distance}{Number of amino acid residues different than reference
using Levenshtein distance}
}
\value{
Seurat or SingleCellExperiment object with epitope information
added to the metadata.
}
\description{
Use meta database of cdr3 sequences as a reference to 
cdr3 aa sequence of the single cells.
}
\examples{
trex_example <- annotateDB(trex_example, 
                        chains = "TRB")
}
