% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cal.data}
\alias{cal.data}
\title{Calibration Measurements}
\format{An object of class \code{data.frame} with 4024 rows and 10 columns.}
\usage{
cal.data
}
\description{
Returns raw calibration experiment data obtained from literature,
 with K values combined with gravimetrically determined sap flux density.
 The data.frame contains 22 studies with 37 different species. The data is used
 within the \code{cal.fd()} function to calculate sap flux density. The data
 originates from the manuscript by Flo \emph{et al.} 2019 and provides a
 description on the genus, species, calibration material, wood porosity and diameter
 of the stem. The presented data is open for public use.
}
\references{
Flo V, Martinez-Vilalta J, Steppe K, Schuldt B, Poyatos, R. 2019.
A synthesis of bias and uncertainty in sap flow methods.
Agricultural and Forest Meteorology 271:362-374 \url{doi: 10.1016/j.agrformet.2019.03.012}

Granier A. 1985. Une nouvelle methode pour la measure du flux de seve brute dans le tronc des arbres.
Annales des Sciences Forestieres 42:193–200 \url{doi: 10.1051/forest:19850204}
}
\keyword{datasets}
