% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{check_token}
\alias{check_token}
\title{Check the validity of a securely-stored API token}
\usage{
check_token(username, api_name = "ldc", keyring_name = NULL, verbose = FALSE)
}
\arguments{
\item{username}{Character string. The username associated with the API account and token being retrieved, typically an email address.}

\item{api_name}{Character string. The abbreviated \(and case-insensitive\) name of the API, e.g. \code{"LDC"}, the token is associated with. Defaults to \code{"ldc"}.}

\item{keyring_name}{Optional character string. The name of the keyring that the password was stored in. If this is \code{NULL} then the keyring matching \code{generate_keyring_name(username)} will be used. Defaults to \code{NULL}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\description{
Attempt to retrieve a previously-stored API token from the user's encrypted keyring.
in order to confirm that it exists and has not yet expired.
}
