% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTrex.R
\name{clonalCommunity}
\alias{clonalCommunity}
\title{Cluster clones using the Trex dimensional reductions}
\usage{
clonalCommunity(
  sc,
  reduction.name = NULL,
  cluster.parameter = KNNGraphParam(k = 30, ...),
  ...
)
}
\arguments{
\item{sc}{Single Cell Object in Seurat or SingleCell Experiment format. In addition, the outputs of distReduction()
and aaReduction() can be used.}

\item{reduction.name}{Name of the dimensional reduction output from runTrex()}

\item{cluster.parameter}{The community detection algorithm in \link[bluster]{bluster}}

\item{...}{For the generic, further arguments to pass to specific methods.}
}
\value{
Single-Cell Object with trex.clusters in the meta.data
}
\description{
Use this to return clusters for clonotypes based on 
the \link[bluster]{bluster} clustering parameters.
}
\examples{
\dontrun{
sc <- clonalCommunity(sc, 
                      reduction.name = NULL, 
                      cluster.parameter = KNNGraphParam())
}
}
