% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{compare_scatter_plot}
\alias{compare_scatter_plot}
\title{compare_scatter_plot}
\usage{
compare_scatter_plot(file1_path, file2_path, output_path, burnin = NULL)
}
\arguments{
\item{file1_path}{path to the saved BIT Gibbs sampling results of input 1.}

\item{file2_path}{path to the saved BIT Gibbs sampling results of input 2.}

\item{output_path}{path to save the barplot.}

\item{burnin}{number of samples used for burn-in. If not specify, BIT will use the half of the iterations as burn in.}
}
\description{
To draw a barplot for the top n TRs.
}
