% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNGAfy.R
\name{CoNGAfy}
\alias{CoNGAfy}
\title{Reduce a single-cell object to representative cells}
\usage{
CoNGAfy(sc, method = "mean", features = NULL)
}
\arguments{
\item{sc}{Single Cell Object in Seurat or Single Cell Experiment format}

\item{method}{"mean" or "dist" Generate a mean value across features by clonotype or 
use the PCA reduction to identify the cell with the minimal euclidean distance from the
clonotype group.}

\item{features}{Selected genes for the reduction DEFAULT: null will return all genes}
}
\value{
Single-cell Object with 1 cell representing 1 clone
}
\description{
Generate a single-cell object that has a representation of
RNA expression by clonotype. This approach was first introduced
in \href{https://pubmed.ncbi.nlm.nih.gov/34426704/}{CoNGA} and was 
adapted to R. Please read and cite the authors work.
}
\examples{
trex_values <- CoNGAfy(trex_example, 
                        method = "mean",
                        features = NULL)
                        
}
