% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reductions.R
\name{distReduction}
\alias{distReduction}
\title{Reduce TCR data based on pairwise edit distance}
\usage{
distReduction(sc, chains = "TRB", edit.method = "lv", n.trim = 0, c.trim = 0)
}
\arguments{
\item{sc}{Single Cell Object in Seurat or SingleCell Experiment format}

\item{chains}{TRA or TRB (no both option)}

\item{edit.method}{distance measures inherited from \link[stringdist]{stringdist} or NULL
to not include edit distance layer in multiplex network}

\item{n.trim}{length from N-terminus to trim on the CDR3 aa sequence}

\item{c.trim}{length from C-terminus to trim on the CDR3 aa sequence}
}
\value{
distance object based on the indicated index measure
}
\description{
Use this to get the pairwise edit distance of the TCRs. Distance methods 
available are from \link[stringdist]{stringdist} package.s
}
\examples{
dist.dim  <- distReduction(trex_example, 
                        chains = "TRB", 
                        edit.method = "lv",
                        n.trim = 0,
                        c.trim = 0)

}
