% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.max.R
\name{dt.max}
\alias{dt.max}
\title{Calculate zero-flow conditions}
\usage{
dt.max(input, methods = c("pd","mw","dr"),
zero.end = 8*60,
zero.start =  1*60,
interpolate = FALSE, det.pd = TRUE,
max.days = 7,
ed.window = 2*60,
vpd.input,
sr.input,
sel.max,
criteria = c(sr = 30, vpd = 0.1, cv = 0.5),
df = FALSE)
}
\arguments{
\item{input}{An \code{is.trex}-compliant object of \eqn{K} values containing
a timestamp and a value column.}

\item{methods}{Character vector of the requested \eqn{\Delta T_{max}}{\Delta Tmax} methods.
Options include “pd” (predawn), “mw” (moving window), “dr” (double regression),
and “ed” (environmental dependent; default= c(“pd”, “mw”, “dr”)).}

\item{zero.end}{Numeric, optionally defines the end of the predawn period.
Values should be in minutes (e.g. predawn conditions until 08:00 = 8 *60).
When not provided, the algorithm will automatically analyse the cyclic behaviour
of the data and define the day length.}

\item{zero.start}{Numeric, optionally defines the beginning of the predawn period.
Values should be in minutes (e.g., 01:00 = 1*60).}

\item{interpolate}{Logical: if \code{TRUE}, detected \eqn{\Delta T_{max}}{\Delta Tmax} values are linearly
interpolated. If \code{FALSE}, constant \eqn{\Delta T_{max}}{\Delta Tmax} values will be selected daily
(default = FALSE).}

\item{det.pd}{Logical; if \code{TRUE} and no zero.end and zero.start values are provided,
predawn \eqn{\Delta T_{max}}{\Delta Tmax} will be determined based on cyclic behaviour of the entire
time-series (default = \code{TRUE}).}

\item{max.days}{Numeric, defines the number of days which the \code{mw} and \code{dr}
methods will consider for their moving-windows.}

\item{ed.window}{Numeric, defines the length of the period considered for assessing the
environmental conditions and stable \eqn{\Delta T_{max}}{\Delta Tmax} values.}

\item{vpd.input}{An \code{is.trex}-compliant object (\code{zoo} time-series or \code{data.frame})
with a timestamp and a value column containing the vapour pressure deficit (\emph{vpd}; in kPa)
with the same temporal extent and time steps as the input data.}

\item{sr.input}{An \code{is.trex}-compliant object (\code{zoo} time-series or \code{data.frame})
with a timestamp and a value column the solar radiation data (\emph{sr}; e.g., global radiation or \emph{PAR})}

\item{sel.max}{Optional \code{zoo} time-series or \code{data.frame} with the specified \eqn{\Delta T_{max}}{\Delta Tmax}.
This is included to change predawn \eqn{\Delta T_{max}}{\Delta Tmax} values selected with the \code{ed} method.}

\item{criteria}{Numeric vector, thresholds for the \code{ed} method.
Thresholds should be provided for all environmental data included in the function
(e.g. \emph{sr} = 30; \emph{vpd} = 0.1; coefficient of variation, \emph{cv} = 0.5)}

\item{df}{Logical; if \code{TRUE}, output is provided in a \code{data.frame}
format with a timestamp and a value (\eqn{\Delta T} or \eqn{\Delta V}) column.
If \code{FALSE}, output is provided as a \code{zoo} object (default = \code{FALSE}).}
}
\value{
A named \code{list} of \code{zoo} time series or \code{data.frame}
objects in the appropriate format for further processing.
List items include:
\describe{
 \item{max.pd}{\eqn{\Delta T_{max}}{\Delta Tmax} time series as determined by the \code{pd} method.}
 \item{max.mw}{\eqn{\Delta T_{max}}{\Delta Tmax} time series as determined by the \code{mw} method.}
 \item{max.dr}{\eqn{\Delta T_{max}}{\Delta Tmax} time series as determined by the \code{dr} method.}
 \item{max.ed}{\eqn{\Delta T_{max}}{\Delta Tmax} time series as determined by the \code{ed} method.}
 \item{daily_max.pd}{daily predawn \eqn{\Delta T_{max}}{\Delta Tmax} as determined by \code{pd}.}
 \item{daily_max.mw}{daily predawn \eqn{\Delta T_{max}}{\Delta Tmax} as determined by \code{mw}.}
 \item{daily_max.dr}{daily predawn \eqn{\Delta T_{max}}{\Delta Tmax} as determined by \code{dr}.}
 \item{daily_max.ed}{daily predawn \eqn{\Delta T_{max}}{\Delta Tmax} as determined by \code{ed}.}
 \item{all.pd}{exact predawn \eqn{\Delta T_{max}}{\Delta Tmax} values detected with \code{pd}.}
 \item{all.pd}{exact predawn \eqn{\Delta T_{max}}{\Delta Tmax} values detected with \code{ed}.}
 \item{input}{\eqn{\Delta T} input data.}
 \item{ed.criteria}{\code{data.frame} of the applied environmental and variability criteria used within \code{ed}.}
 \item{methods}{\code{data.frame} of applied methods to detect \eqn{\Delta T_{max}}{\Delta Tmax}.}
 \item{k.pd}{\eqn{K} values calculated by using the \code{pd} method.}
 \item{k.mw}{\eqn{K} values calculated by using the \code{mw} method.}
 \item{k.dr}{\eqn{K} values calculated by using the \code{dr} method.}
 \item{k.ed}{\eqn{K} values calculated by using the \code{ed} method.}
 }
}
\description{
Determine zero flow conditions (\eqn{\Delta T_{max}}{\Delta Tmax}; or \eqn{\Delta V_{max}}{\Delta Vmax})
according to four methods; namely,
 i) predawn (\code{pd}),
 ii) moving-window (\code{mw}),
 iii) double regression (\code{dr}),
 and iv) environmental dependent (\code{ed}) as applied in Peters et al. 2018.
 The function can provide (\eqn{\Delta T_{max}}{\Delta Tmax} values and subsequent \emph{K} values for all methods.
 All outputs are provided in a list where users can review the provided output and used input data.
}
\examples{
\dontrun{
#perform Delta T maximum calculations
raw   <-
  TREX::is.trex(
    example.data(type = "doy", species = "PCAB"),
    tz = "GMT",
    time.format = "\%H:\%M",
    solar.time = TRUE,
    long.deg = 7.7459,
    ref.add = FALSE
  )
input   <-
  TREX::time_step(
    input = raw,
    start = "2014-05-08 00:00",
    end = "2014-07-25 00:50",
    time.int = 15,
    max.gap = 60,
    decimals = 6,
    df = F
  )
input[which(input < 0.2)] <- NA

vpd_raw   <-
  is.trex(
    vpd,
    tz = "GMT",
    time.format = "(\%m/\%d/\%y \%H:\%M:\%S)",
    solar.time = TRUE,
    long.deg = 7.7459,
    ref.add = FALSE
  )
vpd.input <-
  time_step(
    input = vpd_raw,
    start = "2014-05-08 00:00",
    end = "2014-07-25 00:50",
    time.int = 15,
    max.gap = 60,
    decimals = 6,
    df = F
  )
sr_raw   <-
  is.trex(
    sr,
    tz = "GMT",
    time.format = "(\%m/\%d/\%y \%H:\%M:\%S)",
    solar.time = TRUE,
    long.deg = 7.7459,
    ref.add = FALSE
  )
sr.input <-
  time_step(
    input = sr_raw,
    start = "2014-05-08 00:00",
    end = "2014-07-25 00:50",
    time.int = 15,
    max.gap = 60,
    decimals = 6,
    df = F
  )

output.max <-
  dt.max(
    input,
    methods = c("pd", "mw", "dr", "ed"),
    det.pd = TRUE,
    interpolate = FALSE,
    max.days = 10,
    sr.input = sr.input,
    vpd.input = vpd.input,
    ed.window = 2 * 60,
    criteria = c(sr = 30, vpd = 0.1, cv = 0.5),
    df = FALSE
  )

str(output.max)

plot(output.max$input, ylab = expression(Delta * italic("V")))
lines(output.max$max.pd, col = "green")
lines(output.max$max.mw, col = "blue")
lines(output.max$max.dr, col = "orange")
lines(output.max$max.ed, col = "purple")
points(output.max$all.pd, col = "green", pch = 16)
points(output.max$all.ed, col = "purple", pch = 16)
legend(
  "bottomright",
  c("raw", "max.pd", "max.mw", "max.dr", "max.ed"),
  lty = 1,
  col = c("black", "green", "blue", "orange", "purple"))
}
}
