% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{eTRex}
\alias{eTRex}
\title{eTRex: variational Bayesian Transcriptional Regulators explorer}
\usage{
eTRex(
  file,
  output_path,
  filter_path = NULL,
  show = TRUE,
  format = NULL,
  N = 5000,
  bin_width = 1000,
  genome = c("hg38", "mm10")
)
}
\arguments{
\item{file}{Character. Path to the input file containing genomic regions.}

\item{output_path}{Character. Directory path where results will be saved.}

\item{filter_path}{Character, optional. Path to filter regions file. Default: NULL.}

\item{show}{Logical. Whether to display results tables. Default: TRUE.}

\item{format}{Character, optional. Input file format. Default: NULL.}

\item{N}{Integer. Number of iterations for variational inference. Default: 5000.}

\item{bin_width}{Integer. Width of genomic bins in base pairs. Default: 1000.}

\item{genome}{Character vector. Supported genomes: "hg38", "mm10". Default: c("hg38", "mm10").}
}
\value{
Invisible NULL. Results are saved to the specified output path.
}
\description{
Performs variational Bayesian inference on ChIP-seq data by comparing input regions
with pre-compiled reference data and generating comprehensive analysis results.
}
\examples{
\dontrun{
eTRex(
  file = "hg38_CTCF.bed",
  output_path = "./results",
  N = 3000,
  bin_width = 500
)
}

}
