% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ecosites.r
\name{fetch_edit_ecosites}
\alias{fetch_edit_ecosites}
\title{Fetch ecological sites for a specified MLRA, or all ecological sites if no MLRA is provided}
\usage{
fetch_edit_ecosites(
  mlra = NULL,
  keys = NULL,
  key_type = NULL,
  return_only_id = F,
  key_chunk_size = 100,
  timeout = 60,
  verbose = F
)
}
\description{
Fetch ecological sites for a specified MLRA, or all ecological sites if no MLRA is provided
}
