% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.r
\name{fetch_edit_ecosites}
\alias{fetch_edit_ecosites}
\alias{fetch_edit_community}
\alias{fetch_edit_description}
\title{Fetch ecological sites for a specified MLRA, or all ecological sites if no MLRA is provided}
\usage{
fetch_edit_ecosites(
  mlra = NULL,
  keys = NULL,
  key_type = NULL,
  return_only_id = F,
  key_chunk_size = 100,
  timeout = 60,
  verbose = F
)

fetch_edit_community(
  mlra,
  data_type,
  keys = NULL,
  key_type = NULL,
  query_ecosite = TRUE,
  ecosystem_state_sequence = 1,
  land_use_sequence = 1,
  community_sequence = 1,
  key_chunk_size = 100,
  timeout = 60,
  verbose = F
)

fetch_edit_description(
  mlra,
  data_type,
  keys = NULL,
  key_type = NULL,
  query_ecosite = TRUE,
  key_chunk_size = 100,
  timeout = 60,
  verbose = F
)
}
\description{
Fetch ecological sites for a specified MLRA, or all ecological sites if no MLRA is provided

Fetch rangeland community composition tables

Fetch EDIT descriptions
}
