% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldc.R
\name{fetch_ldc}
\alias{fetch_ldc}
\title{Fetching data from the Landscape Data Commons via API query}
\usage{
fetch_ldc(
  data_type,
  query_parameters = list(),
  username = NULL,
  api_key_name = NULL,
  timeout = 300,
  take = 10000,
  delay = 2000,
  coerce = TRUE,
  base_url = "https://api.landscapedatacommons.org/api/v1/",
  verbose = FALSE,
  keys = deprecated(),
  key_type = deprecated(),
  password = deprecated(),
  key_chunk_size = deprecated(),
  exact_match = deprecated()
)
}
\arguments{
\item{data_type}{Character string. The type of data to query. Note that any variables specified in \code{query_parameters} must appear in the table corresponding to \code{data_type}. To see all valid values, use \code{ldc_table_names()}.}

\item{query_parameters}{List. A list of query parameters which can be recognized by (\code{link[=format_query_parameters]{format_query_parameters()}}) and converted into a query for the API, e.g., \code{list("ProjectKey" = list("equals" = "BLM_AIM"))}. Defaults to an empty list which is equivalent to requesting all available records.}

\item{username}{Optional character string. The username associated with the LDC API account and which has been used to store one or more API keys using (\code{link[=store_api_key]{store_api_key()}}). This is almost certainly an email address. Some data in the Landscape Data Commons are accessible only to users with appropriate credentials. You do not need to supply credentials, but an API request made without them may return fewer or no data. Defaults to \code{NULL}.}

\item{api_key_name}{Optional character string. The name used to store the API key to submit with the query and which is associated with \code{username}. If this is \code{NULL} only data which do not require special permissions will be returned. Defaults to \code{NULL}.}

\item{timeout}{Numeric. The number of seconds to wait for a nonresponse from the API before considering the query to have failed. Defaults to \code{300}.}

\item{take}{Optional numeric. The number of records to retrieve at a time. This is NOT the total number of records that will be retrieved! Queries that retrieve too many records at once can fail, so this allows the process to retrieve them in smaller chunks. The function will keep requesting records in chunks equal to this number until all matching records have been retrieved. If this value is too large (i.e., much greater than about \code{10000}), the server will likely respond with a 500 error. If \code{NULL} then all records will be retrieved in a single pass. Defaults to \code{10000}.}

\item{delay}{Optional numeric. The number of milliseconds to wait between API queries. Querying too quickly can crash an API or get you locked out, so adjust this as needed. Defaults to \code{2000} (2 seconds).}

\item{coerce}{Logical. If \code{TRUE} then the returned values will be coerced into the intended class when they don't match, e.g., if a date variable is a character string instead of a date. Defaults to \code{TRUE}.}

\item{base_url}{Character string. The URL for the API endpoint to use. Defaults to \code{"https://api.landscapedatacommons.org/api/v1/"}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report additional diagnostic messages as it executes. Defaults to \code{FALSE}.}
}
\value{
A data frame of records from the requested \code{data_type} which contain the values from \code{keys} in the variable \code{key_type}.
}
\description{
A function for making API calls to the Landscape Data Commons based on the table, key variable, and key variable values. It will return a table of records of the requested data type from the LDC in which the variable \code{key_type} contains only values found in \code{keys}. See the \href{https://api.landscapedatacommons.org/api-docs}{API documentation} to see which variables (i.e. \code{key_type} values) are valid for each data type.

There are additional functions to simplify querying by spatial location (\code{\link[=fetch_ldc_spatial]{fetch_ldc_spatial()}}) and by ecological site ID (\code{\link[=fetch_ldc_ecosite]{fetch_ldc_ecosite()}}).
}
\examples{
# To retrieve all sampling location metadata collected in the ecological
# sites R036XB006NM and R036XB007NM.
headers <- fetch_ldc(query_parameters = list("EcologicalSiteID" = list("=" = c("R036XB006NM",
                                                                               "R036XB007NM"))),
                                        data_type = "header")
                                        
# To retrieve all Line-Point Intercept collected between 40 and 45 degrees
# latitude (inclusive) and east of -115 degrees longitude (exclusive).
lpi_data <- fetch_ldc(query_parameters = list("Latitude_NAD83" = list(">=" = 40,
                                                                      "<=" = 45),
                                              "Longitude_NAD83" = list(">" = -115)),
                      data_type = "lpi")
                                        
}
\seealso{
\code{\link[=fetch_ldc_spatial]{fetch_ldc_spatial()}} will query for data by spatial location.
\code{\link[=fetch_ldc_ecosite]{fetch_ldc_ecosite()}} will query for data by ecological site ID.
}
