% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldc.R
\name{fetch_ldc_ecosite}
\alias{fetch_ldc_ecosite}
\title{Fetching data from the Landscape Data Commons via API query using ecological site IDs}
\usage{
fetch_ldc_ecosite(
  data_type,
  ecosite_ids,
  username = NULL,
  api_key_name = NULL,
  timeout = 300,
  take = NULL,
  delay = 500,
  base_url = "https://api.landscapedatacommons.org/api/v1/",
  verbose = FALSE,
  keys = deprecated(),
  key_chunk_size = deprecated(),
  exact_match = deprecated()
)
}
\arguments{
\item{data_type}{Character string. The type of data to query. Valid values are: \code{'gap'}, \code{'header'}, \code{'height'}, \code{'lpi'}, \code{'soilstability'}, \code{'speciesinventory'}, \code{'indicators'}, \code{'species'}, \code{'dustdeposition'}, \code{'horizontalflux'}, and \code{'schema'}.}

\item{ecosite_ids}{Character vector. All the ecological site IDs (e.g. \code{"R036XB006NM"}) to search for. The returned data will consist only of records where the designated ecological site ID matched one of these values, but there may be ecological site IDS that return no records.}

\item{username}{Optional character string. The username associated with the LDC API account and which has been used to store one or more API keys using (\code{link[=store_api_key]{store_api_key()}}). This is almost certainly an email address. Some data in the Landscape Data Commons are accessible only to users with appropriate credentials. You do not need to supply credentials, but an API request made without them may return fewer or no data. Defaults to \code{NULL}.}

\item{api_key_name}{Optional character string. The name used to store the API key to submit with the query and which is associated with \code{username}. If this is \code{NULL} only data which do not require special permissions will be returned. Defaults to \code{NULL}.}

\item{timeout}{Numeric. The number of seconds to wait for a nonresponse from the API before considering the query to have failed. Defaults to \code{300}.}

\item{take}{Optional numeric. The number of records to retrieve at a time. This is NOT the total number of records that will be retrieved! Queries that retrieve too many records at once can fail, so this allows the process to retrieve them in smaller chunks. The function will keep requesting records in chunks equal to this number until all matching records have been retrieved. If this value is too large (i.e., much greater than about \code{10000}), the server will likely respond with a 500 error. If \code{NULL} then all records will be retrieved in a single pass. Defaults to \code{NULL}.}

\item{delay}{Optional numeric. The number of milliseconds to wait between API queries. Querying too quickly can crash an API or get you locked out, so adjust this as needed. Defaults to \code{500}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report additional diagnostic messages as it executes. Defaults to \code{FALSE}.}
}
\value{
A data frame of records from the requested \code{data_type} which contain the values from \code{keys} in the variable \code{key_type}.
}
\description{
This is a wrapper for \code{\link[=fetch_ldc]{fetch_ldc()}} which streamlines retrieving data by ecological site IDs. Most tables in the LDC do not include ecological site information and so this function will identify the PrimaryKeys associated with the requested ecological site(s) and retrieve the requested data associated with those PrimaryKeys.

When you already know the associated PrimaryKeys, use \code{\link[=fetch_ldc]{fetch_ldc()}} instead. If you want to retrieve data associated with polygons, use \code{\link[=fetch_ldc_spatial]{fetch_ldc_spatial()}}.
}
\examples{
# To retrieve all LPI records associated with the ecological sites R036XB006NM and R036XB007NM
fetch_ldc_ecosite(data_type = "lpi",
                  ecosite_ids = c("R036XB006NM",
                                  "R036XB007NM"))
}
\seealso{
\code{\link[=fetch_ldc]{fetch_ldc()}} will query for data by any key values.
\code{\link[=fetch_ldc_spatial]{fetch_ldc_spatial()}} will query for data by spatial location.
}
