% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldc.R
\name{fetch_ldc_metadata}
\alias{fetch_ldc_metadata}
\title{Fetching data from the Landscape Data Commons via API query}
\usage{
fetch_ldc_metadata(
  data_type,
  timeout = 300,
  delay = 500,
  base_url = "https://api.landscapedatacommons.org/api/v1",
  verbose = FALSE
)
}
\arguments{
\item{data_type}{Character string. The type of data to query. To see all valid values, use \code{ldc_table_names()}.}

\item{timeout}{Numeric. The number of seconds to wait for a nonresponse from the API before considering the query to have failed. Defaults to \code{300}.}

\item{delay}{Optional numeric. The number of milliseconds to wait between API queries. Querying too quickly can cause issues, so adjust this as needed. Defaults to \code{500}.}

\item{base_url}{Character string. The URL for the API endpoint to use. Defaults to \code{"https://api.landscapedatacommons.org/api/v1/"}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report additional diagnostic messages as it executes. Defaults to \code{FALSE}.}
}
\value{
A data frame of metadata for the requested table. The data frame will contain all variables returned from the server plus an additional character variable called \code{"data_class_r"} which contains the R equivalent to the class of the variable called \code{"data_type"}.
}
\description{
A wrapper for (\code{\link[=fetch_ldc]{fetch_ldc()}}) which will return metadata for the specified table.
}
