% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{fetch_rap}
\alias{fetch_rap}
\title{Get data from the Rangeland Analysis Platform}
\usage{
fetch_rap(polygons, data_type = "cover", mask = TRUE, year = NULL)
}
\arguments{
\item{polygons}{Polygon sf object. The area to return data for.}

\item{mask}{Logical. If \code{TRUE} then the request to the API will return data only about areas NOT identified as cropland, water, or human development. Defaults to \code{TRUE}.}

\item{year}{Optional numeric. If this corresponds to a year for which there are RAP data (1986 through the last completed calendar year) then the query will return only data for that year. If it's \code{NULL} then data for every available year will be returned. Defaults to \code{NULL}.}

\item{data_type.}{Character string. The type of data to request. Valid values are \code{"cover"}, \code{"covermeterology"}, \code{"production"}, and \code{"production16day"}. Defaults to \code{"cover"}.}
}
\value{
A data frame containing the requested data for the provided polygon.
}
\description{
Query the RAP for tabular cover, meterological, or production data for a given polygon.
}
