% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{filter_peaks}
\alias{filter_peaks}
\title{function used to filter converted peaks based on distal elements}
\usage{
filter_peaks(input_vec, filter_vec)
}
\arguments{
\item{input_vec}{vectorized object}
}
\value{
index of peak index after the filtering
}
\description{
function used to filter converted peaks based on distal elements
}
