% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldc.R
\name{format_query_parameters}
\alias{format_query_parameters}
\title{Format query parameters for use with the LDC API using POST}
\usage{
format_query_parameters(...)
}
\arguments{
\item{...}{The parameters for the query. These can be provided in three different ways:
1) NAMED arguments where the names are the variables in the data table to use for the query parameters and the values of the arguments are named lists. The lists must be formatted so that the names within the list are the operators to use and the values are vectors of the values associated with each operator. For example, \code{"Latitude" = list(">=" = 40, "<" = 50)} will produce an output to request data where the variable Latitude contains a value greater than or equal to 40 and less than 50.
2) A list of named lists formatted as previously described. The names of the lists must be the names of the variables that they apply to, e.g. \code{list("Longitude" = list(">=" = -105, "<=" = -100), "Latitude" = list("gt" = 40, "lt" = 50))}. This is to make it easier to programmatically generate queries.
3) Parameter strings as produced by \code{stringify_query_parameter()}. If using options 1 or 2, those lists will be run through \code{stringify_query_parameter()}.

Valid operators (and therefore names within lists) are:
* \code{"!="} to exclude one or more values from the query results, e.g., \code{"Source" = list("!=" = "BLM_AIM")} will prevent records with "BLM_AIM" in the Source variable from being included in query results.
* \code{"="} to limit query results to where only the specified value or values occur, e.g. \code{"Source" = list("=" = c("BLM_AIM", "LMF"))} will prevent records with any value other than "BLM_AIM" or "LMF" in the Source variable from being included in query results.
* \code{">"} to limit query results to only where values greater than the provided number occur, e.g. \code{"Latitude_NAD83" = list(">" = 40)} will prevent records with any value less than or equal to 40 in the Longitude variable from being included in query results.
* \code{">="} to limit query results to only where values greater than or equal to the provided number occur, e.g. \code{"Latitude_NAD83" = list(">=" = 40)} will prevent records with any value less than 40 in the Longitude variable from being included in query results.
* \code{"<"} to limit query results to only where values less than the provided number occur, e.g. \code{"Latitude_NAD83" = list(">" = 40)} will prevent records with any value greater than or equal to 40 in the Longitude variable from being included in query results.
* \code{"<="} to limit query results to only where values less than the provided number occur, e.g. \code{"Latitude_NAD83" = list(">=" = 40)} will prevent records with any value greater than 40 in the Longitude variable from being included in query results.

See (\code{\link[=ldc_api_operators]{ldc_api_operators()}}) for additional recognized operator aliases.

A list may contain multiple operators. For example, \code{"Longitude_NAD83" = list(">" = -105, "<" = -100)} will limit query results to only records where the Latitude value is between -105 and -100.}
}
\value{
A character string suitable for use as the body of a POST submission to the LDC API.
}
\description{
This converts specially-formatted lists (or already correctly-formatted strings) into a single character string suitable for use as body of a POST submission to the LDC API. If a variable appears more than once in the inputs, it will be combined in the output. Currently, this does not support the use of AND or OR with parameters and so all parameters will be assessed using AND. In the case of conflicting parameters (i.e., using "equals" with any other operator for the same variable) impossible operators will be dropped.
}
\examples{
# To produce a character string that can be used as the body for a query which
# will return only records with Latitude_NAD83 values between 40 and 45 (inclusive)
# and Longitude_NAD83 values greater than -115
query_parameters <- list("Latitude_NAD83" = list(">=" = 40,
                                                 "<=" = 45),
                         "Longitude_NAD83" = list(">" = -115))
body <- format_query_parameters(query_parameters)

# This is equivalent to above
body_alternate <- format_query_parameters("Latitude_NAD83" = list(">=" = 40,
                                                 "<=" = 45),
                                          "Longitude_NAD83" = list(">" = -115))
identical(body, body_alternate)

}
\seealso{
[ldc_api_operators()], [stringify_query_parameters()]
}
