% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{generate_keyring_name}
\alias{generate_keyring_name}
\title{Create a standardized name for a trex keyring based on an API username.}
\usage{
generate_keyring_name(username)
}
\arguments{
\item{username}{Character string. The username associated with the LDC API account. This is almost certainly an email address.}
}
\value{
A character string matching the standard pattern for a trex keyring name.
}
\description{
The format is the provided username stripped of any domain plus "_trex_keyring".
}
