% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_stored_key}
\alias{get_stored_key}
\title{Retrieve a securely-stored API key}
\usage{
get_stored_key(
  username,
  api_key_name = NULL,
  keyring_name = NULL,
  accept_failure = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{username}{Character string. The username associated with the API key being stored, typically an email address.}

\item{api_key_name}{Optional character string. The name that the API key was stored under in the keyring. Defaults to \code{NULL} which will attempt to retrieve a key called \code{"default"}.}

\item{keyring_name}{Optional character string. The name of the keyring that the API key was stored in. If this is \code{NULL} then the keyring matching \code{generate_keyring_name(username)} will be used. Defaults to \code{NULL}.}

\item{accept_failure}{Logical. If \code{TRUE} then the function will return \code{NULL} if a key can't be found. Otherwise an error message will be produced. Defaults to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\description{
Retrieve a previously-stored API key from the user's encrypted keyring.
There is no argument for the keyring password and the user must provide it when interactively prompted when this function is run and the keyring is locked.
The keyring is managed by the user's OS and will remain unlocked for a period of time controlled by the OS, so the user will not always be prompted for a password.
Other trex functions for requesting data from
an API will use their own \code{username} and \code{api_key_name} to access keys
stored in this way.
}
