% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_stored_token}
\alias{get_stored_token}
\title{Retrieve a securely-stored API token}
\usage{
get_stored_token(
  username,
  api_name = "ldc",
  keyring_name = NULL,
  accept_failure = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{username}{Character string. The username associated with the API account and token being retrieved, typically an email address.}

\item{api_name}{Character string. The abbreviated \(and case-insensitive\) name of the API, e.g. \code{"LDC"}, the token is associated with. Defaults to \code{"ldc"}.}

\item{keyring_name}{Optional character string. The name of the keyring that the password was stored in. If this is \code{NULL} then the keyring matching \code{generate_keyring_name(username)} will be used. Defaults to \code{NULL}.}

\item{accept_failure}{Logical. If \code{TRUE} then the function will return \code{NULL} if a token can't be found. Otherwise an error message will be produced. Defaults to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\description{
Retrieve a previously-stored API token from the user's encrypted keyring.
}
