% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_trex_keyring_name}
\alias{get_trex_keyring_name}
\title{Find the keyring associated with the provided username.}
\usage{
get_trex_keyring_name(username, verbose = FALSE)
}
\arguments{
\item{username}{Character string. The username associated with the LDC API account. This is almost certainly an email address.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
A character string matching the standard pattern for a trex keyring name if that keyring exists OR \code{NULL} in cases where keyring does not exist or the system does not support multiple keyrings.
}
\description{
If the system doesn't support multiple keyrings, this returns \code{NULL} which is
interpreted by keyring:: functions as the default system keyring.
If the system DOES support multiple keyrings, this will check for the
standard keyring name produced by \code{generate_keyring_name()} and return that name if the keyring exists or \code{NULL} if it doesn't.
}
