% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldc.R
\name{ldc_api_operators}
\alias{ldc_api_operators}
\title{Look up LDC operators}
\usage{
ldc_api_operators(operator = NULL)
}
\arguments{
\item{operator}{Optional character string. A user-supplied name for an operator, e.g. \code{"!="} or \code{"greaterthan"}. This is case-insensitive. If \code{NULL} then the lookup table of recognized operators and their aliases will be returned instead. Defaults to \code{NULL}.}
}
\value{
Either a data frame of operators and their recognized aliases OR an operator as a character string.
}
\description{
Either get a data frame of operators recognized by the LDC API or the operator
that corresponds to a recognized human-friendly alias.
}
\examples{
# Get the complete lookup table.
ldc_api_operator_lookup <- ldc_api_operators()

# Get the operator associated with an alias.
ldc_api_operators(operator = "oneof")
ldc_api_operators(operator = ">=")
}
