% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldc.R
\name{ldc_table_names}
\alias{ldc_table_names}
\title{Look up LDC table names}
\usage{
ldc_table_names(alias = NULL)
}
\arguments{
\item{alias}{Optional character string. A user-supplied name for a data type, e.g. \code{"gap"}. This is case-insensitive and will ignore whitespace and nonalphabetic characters. If \code{NULL} then the lookup table of LDC table names and recognized aliases will be returned instead. Defaults to \code{NULL}.}
}
\value{
Either a data frame of LDC table names and their recognized aliases OR the name of an LDC table as a character string.
}
\description{
Either get a data frame of tables in the LDC which can be queried or get the
name of a table in the LDC that corresponds to a recognized human-friendly
alias.
}
\examples{
# Get the complete lookup table.
ldc_table_lookup <- ldc_table_names()

# Get the name of the table associated with gap data.
ldc_table_lookup <- ldc_table_names(alias = "gap")
}
