% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{logistic}
\alias{logistic}
\title{function used to compute the logistic transformation}
\usage{
logistic(x)
}
\arguments{
\item{x}{vectorized object}
}
\value{
None
}
\description{
function used to compute the logistic transformation
}
