% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{outlier}
\alias{outlier}
\title{Automated and manual outlier detection}
\usage{
outlier()
}
\value{
Runs an interactive shiny app with option to save output as a \code{.Rda} file
}
\description{
This interactive app allows filtering outlier data via a specified threshold
and by manually clicking on problematic observations. The data must be provided as a
\code{data.frame} saved to disk in \code{.Rda} format.
}
\examples{
\dontrun{
# find example file path
system.file("exdata", "example.RData", package = "TREX", mustWork = TRUE)
# either copy-paste this into the navigation bar of the file selection window
# or navigate here manually for selection
outlier()

}

}
