% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{preci}
\alias{preci}
\title{Daily precipitation (raw)}
\format{An object of class \code{data.frame} with 3613 rows and 2 columns.}
\usage{
preci
}
\description{
Returns raw precipitation measurements in mm day-1, collected at ORIGIN
}
\keyword{datasets}
