% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometry_pred.R
\name{predict_allometry}
\alias{predict_allometry}
\title{Predict allometric parameters from DBH}
\usage{
predict_allometry(
  input_dbh_cm,
  parameter = c("Bark_cm", "Sap_cm", "Saparea_cm2"),
  species = NULL,
  ...
)
}
\arguments{
\item{input_dbh_cm}{numeric, single value or vector of DBH measurements in cm}

\item{parameter}{character, one or multiple of \code{"Bark_cm"}, \code{"Sap_cm"}, \code{"Saparea_cm2"}}

\item{species}{character, one or multiple of the species given in \code{\link{allometry.data}}.
Defaults to \code{NULL}, selecting all species.}

\item{...}{additional arguments passed to \code{predict()}}
}
\value{
A data.frame containing predictions for each provided parameter and species
}
\description{
Predict allometric parameters from DBH
}
\examples{
  predict_allometry(input_dbh_cm = 10:15)

  predict_allometry(input_dbh_cm = 10:15,
    parameter = "Bark_cm",
    species   = c("Pinus banksiana",
                "Betula alleghaniensis"))
}
