% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldc.R
\name{query_ldc}
\alias{query_ldc}
\title{Submit a POST query to the LDC API}
\usage{
query_ldc(
  data_type,
  body = NULL,
  api_key = NULL,
  token = NULL,
  timeout = 300,
  base_url = "https://api.landscapedatacommons.org/api/v1/",
  verbose = FALSE
)
}
\arguments{
\item{data_type}{Character string. The type of data to query. To see all valid values, use \code{ldc_table_names()}.}

\item{body}{Optional list or character string. This is passed to (\code{link[=format_query_parameters]{format_query_parameters()}}) before being included as the body of the POST submission to the API. If this is \code{NULL} then query will return all available records. Defaults to \code{NULL}.}

\item{api_key}{OPTIONAL character string. The API key to submit with the request. It is very strongly recommended that you never, ever use your plaintext API key as the value here and instead use (\code{link[=get_stored_key]{get_stored_key()}}) to pull a previously-stored, encrypted key. If \code{NULL} then no key will be used and only data which do not require special permissions will be returned. Defaults to \code{NULL}.}

\item{timeout}{Numeric. The number of seconds to wait for a nonresponse from the API before considering the query to have failed. Defaults to \code{300}.}

\item{base_url}{Character string. The URL for the API endpoint to use. Defaults to \code{"https://api.landscapedatacommons.org/api/v1/"}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report additional diagnostic messages as it executes. Defaults to \code{FALSE}.}
}
\value{
A data frame containing the response from the API.
}
\description{
Given a body string (or parsable list of query parameters) and a target table,
return all matching records available with the permissions associated with any
provided credentials.

This is the function used by all functions with names that start with "fetch_ldc".
}
\seealso{
[fetch_ldc()], [fetch_ldc_spatial()], [fetch_ldc_ecosite()], and [fetch_ldc_metadata()]
}
