% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTrex.R
\name{quietTCRgenes}
\alias{quietTCRgenes}
\title{Remove TCR genes from variable gene results}
\usage{
quietTCRgenes(sc, assay = NULL)
}
\arguments{
\item{sc}{Single-cell object in Seurat format or vector of variable genes to use in reduction}

\item{assay}{The Seurat assay slot to use to remove TCR genes from, NULL value will default to
the default assay}
}
\value{
Seurat object or vector list with TCR genes removed.
}
\description{
Most single-cell workflows use highly-expressed and highly-variable
genes for the initial calculation of PCA and subsequent dimensional
reduction. This function will remove the TCR genes from the variable
features in the Seurat object or from a vector generated by
the Bioconductor scran workflow.
}
\examples{
x <- trex_example
x <- quietTCRgenes(x)

}
\author{
Nicky de Vrij Nick Borcherding
}
