% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{rank_plot}
\alias{rank_plot}
\title{rank_plot}
\usage{
rank_plot(
  file_path = NULL,
  output_path,
  burnin = NULL,
  n = 10,
  colors = "NPG",
  main = NULL,
  xlab = "BIT score",
  ylab = "TR symbols"
)
}
\arguments{
\item{file_path}{path to the saved BIT Gibbs sampling results.}

\item{output_path}{path to save the barplot.}

\item{burnin}{number of samples used for burn-in. If not specify, BIT will use the half of the iterations as burn in.}

\item{n}{top n TRs will show in the barplot, default: 10.}

\item{colors}{colors for each bar, default "NPG" for n<=10, has to be manually specified if n>10.}

\item{main}{main title for the barplot, default: NULL.}

\item{xlab}{x axis label, default: BIT score.}

\item{ylab}{y axis label, default: TR symbols.}
}
\value{
a data.frame object contains TR names, theta_i, BIT scores and 95 CIs for each TR.
}
\description{
To draw a barplot for the top n TRs.
}
