% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTrex.R
\name{runTrex}
\alias{runTrex}
\title{Trex single cell calculation}
\usage{
runTrex(
  sc,
  chains = "TRA",
  method = "encoder",
  encoder.model = "VAE",
  encoder.input = "AF",
  theta = pi,
  reduction.name = "Trex"
)
}
\arguments{
\item{sc}{Single Cell Object in Seurat or SingleCell Experiment format}

\item{chains}{TRA or TRB}

\item{method}{"encoder" = using deep learning autoencoders or 
"geometric" = geomteric transformations based on BLOSUM62 matrix}

\item{encoder.model}{"AE" = dense autoencoder or "VAE" = variation autoencoder}

\item{encoder.input}{Atchley factors (\strong{AF}), Kidera factors ((\strong{AF})),
or or One-Hot Encoder (\strong{OHE}).}

\item{theta}{angle to use for geometric transformation}

\item{reduction.name}{Keyword to save Trex reduction. Useful if you want
to try Trex with multiple parameters}
}
\value{
Seurat or SingleCellExperiment object with Trex dimensions placed 
into the dimensional reduction slot.
}
\description{
Run Trex algorithm with Seurat or SingleCellExperiment pipelines
}
\examples{
trex_example <- runTrex(trex_example, 
                        chains = "TRA",
                        method = "encoder",
                        encoder.model = "VAE",
                        encoder.input = "AF")
                        
}
