% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{setup_keyring}
\alias{setup_keyring}
\title{Establish a user-specific trex keyring if possible}
\usage{
setup_keyring(
  username = NULL,
  keyring_password = NULL,
  interactive = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{username}{Character string. The username associated with the LDC API account. This is almost certainly an email address.}

\item{keyring_password}{OPTIONAL (and discouraged) character string. The password to use to encrypt the keyring and which will be used to access and API keys stored in the keyring in the future. It is very strongly recommended that you never save a plaintext password in your code. If this is \code{NULL} and \code{interactive} is \code{TRUE} then you will be prompted to type a password when the function is run. Defaults to \code{NULL}.}

\item{interactive}{Logical. If \code{TRUE} and either \code{username} or \code{keyring_password} is \code{NULL} then interactive prompts will ask you for the missing values. Defaults to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\description{
Attempt to create a keyring for the user to store their API keys in. The keyring name is created using \code{generate_keyring_name()}.
}
