% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sr}
\alias{sr}
\title{Solar radiation measurements (raw)}
\format{An object of class \code{data.frame} with 318334 rows and 2 columns.}
\usage{
sr
}
\description{
Returns raw solar radiation measurements in UNITS, collected at ORIGIN
}
\keyword{datasets}
