% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{store_api_token}
\alias{store_api_token}
\title{Securely store an API token for future use}
\usage{
store_api_token(
  username,
  api_name = "ldc",
  keyring_name = NULL,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{username}{Character string. The username associated with the API account being used to retrive and store the token. Typically an email address.}

\item{api_name}{Character string. The name of the API associated with the password. Defaults to \code{"ldc"}.}

\item{keyring_name}{Optional character string. The name of the keyring that the API token will be stored in. If this is \code{NULL} then the keyring matching \code{generate_keyring_name(username)} will be used. Defaults to \code{NULL}.}

\item{overwrite}{Logical. If \code{TRUE} then an existing token stored for the provided username for the API will be overwritten. Defaults to \code{FALSE}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\description{
Store an API token in the user's encrypted keyring. The token is retrieved
at the time of the function call and cannot be provided as an argument. An
existing stored token will only be overwritten if it has expired or the
argument \code{overwrite} is \code{TRUE}.
}
