% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldc.R
\name{stringify_query_parameter}
\alias{stringify_query_parameter}
\title{Create an API-parsable character string for a parameter}
\usage{
stringify_query_parameter(variable, operator = "equals", values)
}
\arguments{
\item{variable}{Character string. The name of the variable in the target table which the operator will be applied to.}

\item{operator}{Character string. The name of the operator recognized by the LDC API or a valid alias, e.g. \code{">="}. The function (\code{\link[=ldc_api_operators]{ldc_api_operators()}}) can be used to see all recognized values for this argument. Defaults to \code{"equals"}.}

\item{values}{Single character string or numeric value OR a vector of character strings or numeric values. The value or values to apply the operator to. In the case of any operators other than \code{"equals"} or \code{"notequalto"} \(or an equivalent alias\) this must be numeric and a single value rather than a vector.}
}
\value{
A character string suitable for inclusion in the body of a POST submission to the LDC API.
}
\description{
Parameters submitted to the LDC API must be done via specifically-formatted
character strings. This will produce a character string for a parameter.
}
\examples{
# Produce a string that will limit the API's returned results to only records
# where ProjectKey is "BLM_AIM" or "NWERN"
stringify_query_parameter(variable = "ProjectKey",
                          values = c("BLM_AIM", "NWERN"))
stringify_query_parameter(variable = "ProjectKey",
                          operator = "oneof",
                          values = c("BLM_AIM", "NWERN"))
stringify_query_parameter(variable = "ProjectKey",
                          operator = "in",
                          values = c("BLM_AIM", "NWERN"))
                          
# Produce a string that will limit the API's returned results to only records
# where Latitude_NAD83 is greater than or equal to 40
stringify_query_parameter(variable = "Latitude_NAD83",
                          operator = ">="
                          values = 40)
stringify_query_parameter(variable = "Latitude_NAD83",
                          operator = "gte"
                          values = 40)

}
