% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trex-package.R
\docType{package}
\name{trex-package}
\alias{trex}
\alias{trex-package}
\title{trex: Tools for accessing Landscape Data Commons (LDC) and Ecosystems Dynamics Interpretive Tool (EDIT) APIs}
\description{
Tools for accessing Landscape Data Commons (LDC), Ecosystems Dynamics Interpretive Tool (EDIT), and Rangeland Analysis Platform (RAP) APIs
}
\keyword{internal}
