% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trex.Data.R
\docType{data}
\name{Trex.Data}
\alias{Trex.Data}
\title{Amino acid properties and ranges for
the autoencoder models along tensor}
\description{
Amino acid properties and ranges for
the autoencoder models along tensor
}
