% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trex.database.R
\docType{data}
\name{Trex.database}
\alias{Trex.database}
\title{Database for TCR-antigen relationships
for both TRA and TRB chain}
\description{
Database for TCR-antigen relationships
for both TRA and TRB chain
}
