% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trex_example.R
\docType{data}
\name{trex_example}
\alias{trex_example}
\title{A seurat object of 3200 single T cells derived
a CD3+ sorted severe COVID-19 patient}
\description{
A seurat object of 3200 single T cells derived
a CD3+ sorted severe COVID-19 patient
}
