% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vpd}
\alias{vpd}
\title{Vapor pressure deficit measurements (raw)}
\format{An object of class \code{data.frame} with 318429 rows and 2 columns.}
\usage{
vpd
}
\description{
Returns raw vapor pressure deficit measurements in UNITS, collected at ORIGIN
}
\keyword{datasets}
