% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{alignment_wrapper}
\alias{alignment_wrapper}
\title{Count the 'good' and 'informative' cases by comparing input with the reference database.}
\usage{
alignment_wrapper(input_vec, bin_width, genome = c("hg38", "mm10"))
}
\arguments{
\item{input_vec}{A numeric vector containing indices of transformed regions
obtained by applying \code{import_input_regions}.}

\item{bin_width}{Width of bin in base pairs.}

\item{genome}{The genome of TR ChIP-seq data, either "hg38" or "mm10".}
}
\value{
A data frame with three columns: TR labels, number of 'good' cases,
  and number of 'total' informative cases.
}
\description{
Count the 'good' and 'informative' cases by comparing input with the reference database.
}
