% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.data.R
\name{example.data}
\alias{example.data}
\title{Generate example TDM input data}
\usage{
example.data(type = "timestamp")
}
\arguments{
\item{type}{Character string, indicating whether the example data should be
displayed with a timestamp (default = “timestamp”)
or whether a separate day of day of year and hour column
are provided (“doy”).}
}
\value{
A data.frame containing TDM measurements according to a specific type.
}
\description{
This returns a data.frame containing standard TDM (thermal dissipation method)
 measurements provided in two different formats. The data is obtain from \code{tdm.data}
 where \eqn{\Delta V} measurements are provide for i) Norway spruce (\emph{Picea abies} Karts.)
 growing in a valley in the Swiss Alps or ii) black spruce (\emph{Picea mariana} Mill. BSP)/eastern
 larch (\emph{Larix laricina} Du Roi K. Koch) growing at the southern limit of the boreal
 ecozone in central Canada.
}
\details{
This dataset can be applied for testing the functions provided in the package.
}
\examples{

# get example data
input_data <- example.data(type = "timestamp")
input_data <- example.data(type = "doy")
head(input_data)

}
