% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ecosites.r
\name{fetch_ecosites}
\alias{fetch_ecosites}
\title{Fetch ecological sites for a specified MLRA, or all ecological sites if no MLRA is provided}
\usage{
fetch_ecosites(mlra = NULL)
}
\description{
Fetch ecological sites for a specified MLRA, or all ecological sites if no MLRA is provided
}
