% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.r
\name{fetch_edit}
\alias{fetch_edit}
\alias{fetch_mlra_codes}
\title{Fetch EDIT data}
\usage{
fetch_edit(
  mlra = NULL,
  ecosite = NULL,
  data_type,
  tall = TRUE,
  keys = NULL,
  key_type = NULL,
  key_chunk_size = 100,
  timeout = 60,
  delay = 500,
  verbose = FALSE,
  path_unparsable_data = NULL
)

fetch_mlra_codes(verbose = FALSE)
}
\arguments{
\item{mlra}{Character string, vector of character strings, or in rare 
instances NULL. The Major Land Resource Area (MLRA) or MLRAs to query. Only 
records from these MLRAs will be returned. Exact MLRA codes are required, to 
see a list of all MLRAs and ecological sites use fetch_mlra_codes().}

\item{data_type}{Restricted character string. The shorthand name of the type 
of data to retrieve. Must be one of: 'mlra', 'ecosites', 'climate', 
'landforms', physiography interval', 'physiography nominal', 'physiography 
ordinal', annual production', 'overstory', 'understory', 'rangeland', 
'surface cover', parent material', 'soil interval', 'soil nominal', 'soil 
ordinal', soil profile', 'texture', 'state narratives', or 'transition 
narratives'.

For information on what these tables contain, see 
https://edit.jornada.nmsu.edu/resources/esd, under tab-delimited data.}

\item{tall}{Optional logical. The function will output tall/long data if 
\code{TRUE}, otherwise data will be in wide format. Defaults to \code{TRUE}.}

\item{keys}{Optional character vector. A character vector of all the values 
to search for in \code{key_type}. The returned data will consist only of 
records where \code{key_type} contained one of the key values, but there may 
be keys that return no records. If \code{NULL} then the entire table will be 
returned. Defaults to \code{NULL}.}

\item{key_type}{Optional character string. Variable to query using 
\code{keys}. Must be of of: 'precipitation', 'frost free days', 'elevation', 
'slope', 'landform', 'parent material origin', 'parent material kind', or 
'surface texture.' Defaults to \code{NULL}.}

\item{key_chunk_size}{Optional numeric. The number of keys to send in a 
single query. Very long queries fail, so the keys may be chunked into smaller 
queries with the results of all the queries being combined into a single 
output. Defaults to \code{100}.}

\item{timeout}{Optional numeric. The number of seconds to wait for a 
nonresponse from the API before considering the query to have failed. 
Defaults to \code{60}.}

\item{delay}{Optional numeric. The number of milliseconds to wait between 
API queries. Querying too quickly can crash an API or get you locked out, so 
adjust this as needed. Defaults to \code{500}.}

\item{verbose}{Optional logical. If \code{TRUE} then the function will report 
additional diagnostic messages as it executes. Defaults to \code{FALSE}.}

\item{path_unparsable_data}{Optional character string. Because the delimiting 
character is sometimes entered into narrative text data, data from certain 
ecological sites are not parsable and must be excluded from function output. 
If provided, then the function will save unparsable data to a folder 
specified here for manual processing. Defaults to \code{NULL}, meaning 
unparsable data is not saved.}
}
\value{
A data frame with the requested EDIT data.
}
\description{
Fetch EDIT data
}
\examples{
# To retrieve ecological sites from MLRA 039X 
fetch_edit(mlra = "042B", data_type = "ecosites")
# To retrieve ecological sites from MLRAs 039X and 040X
fetch_edit(mlra = c("042B", "042C"), data_type = "ecosites")
# To retrieve climatic feature descriptions from all ecological sites in 
# MLRAs 042V and 042C
fetch_edit(mlra = c("042B", "042C"), data_type = "climate")
# To retrieve climatic feature descriptions from ecological sites that exist 
# with slope between 15 and 30\%, from MLRAs 042B and 042C
# Note: this includes all sites where the key variable range overlaps with the given 
# range. For example this will return sites with slope range 10-30\%.
fetch_edit(mlra = c("042B", "042C"), data_type = "climate", keys = "5:15", key_type = "slope")
# Data defaults ot a tall format. To retrieve wide climate data from MLRAs 
# 042B and 042C with slope range 5-15\%
fetch_edit(mlra = c("042B", "042C"), data_type = "climate", keys = "5:15", key_type = "slope", tall = FALSE)
# MLRA codes must be exact. To see all MLRA codes and names
fetch_mlra_codes()

}
