% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_api.R
\name{fetch_ldc}
\alias{fetch_ldc}
\title{Query the Landscape Data Commons}
\usage{
fetch_ldc(keys, key_type = "ecosite", data_type = "lpi", verbose = FALSE)
}
\arguments{
\item{keys}{Character vector. The character strings containing the values to match in the query/queries.}

\item{key_type}{Character string. The type of key values being used. This determines which variable in the Landscape Data Commons to filter using the values \code{keys}. Must be \code{"ecosite"} for \code{EcologicalSiteId}, \code{"primarykey"} for \code{PrimaryKey}, or \code{"projectkey"} for \code{ProjectKey}. Defaults to \code{"ecosite"}.}

\item{data_type}{Character string. Determines which table to download data from. Valid values are \code{"lpi"}, \code{"height"}, \code{"gap"}, \code{"soil"}, \code{"species"}, \code{"speciesinventory"}, and \code{"header"}. Defaults to \code{"lpi"}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report with diagnostic messages as it runs. Defaults to \code{FALSE}.}
}
\description{
Fetch data from the Landscape Data commons, querying by ecological site ID, PrimaryKey, or ProjectKey.
}
