% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{import_input_regions}
\alias{import_input_regions}
\title{Transform the input file to vector}
\usage{
import_input_regions(file, format = NULL, bin_width = 1000)
}
\arguments{
\item{file}{file path to the user-input.}

\item{format}{format can be .bed, .narrowPeak, .broadPeak and .bigNarrowPeak.}

\item{bin_width}{desired width of bin, should be in 100/500/1000.}
}
\value{
A numeric vector contains the index of peaks with pre-specified number of bins in each chromosome.
}
\description{
Transform the input file to vector
}
