% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTrex.R
\name{maTrex}
\alias{maTrex}
\title{Main Trex interface}
\usage{
maTrex(sc, chains = "TRB", AA.properties = "AF")
}
\arguments{
\item{sc}{Single Cell Object in Seurat or SingleCell Experiment format}

\item{chains}{TRA or TRB}

\item{AA.properties}{Amino acid properties to use for distance calculation: 
"AF" = Atchley factors, "KF" = Kidera factors, "both" = AF and KF.}
}
\value{
Trex encoded values from the autoencoder
}
\description{
Use this to run the Trex algorithm to return latent vectors in 
the form of a matrix or if you prefer, a maTrex.
}
\examples{
trex_values <- maTrex(trex_example, 
                        chains = "TRA",
                        AA.properties = "AF")
                        
}
