% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTrex.R
\name{maTrex}
\alias{maTrex}
\title{Main Trex interface}
\usage{
maTrex(
  sc,
  chains = "both",
  edit.method = "lv",
  AA.properties = c("AF", "KF", "both"),
  AA.method = "auto",
  c.trim = 0,
  n.trim = 0,
  nearest.method = "threshold",
  threshold = 0.85,
  near.neighbor = NULL,
  add.INKT = TRUE,
  add.MAIT = TRUE,
  n.dim = 30,
  species = "human"
)
}
\arguments{
\item{sc}{Single Cell Object in Seurat or SingleCell Experiment format}

\item{chains}{TRA, TRB, both}

\item{edit.method}{distance measures inherited from \link[stringdist]{sstringdist} or NULL
to not include edit distance layer in multiplex network}

\item{AA.properties}{Amino acid properties to use for distance calculation: 
"AF" = Atchley factors, "KF" = Kidera factors, "both" = AF and KF, or NULL to n
ot include edit distance layer in multiplex network}

\item{AA.method}{method for deriving pairwise distances from amino acid properties, 
either "mean" or the average amino acid characteristics for given cdr3 sequence or "auto"
for using an auto-encoder model.}

\item{c.trim}{length from C-terminus to trim on the CDR3 aa sequence}

\item{n.trim}{length from N-terminus to trim on the CDR3 aa sequence}

\item{nearest.method}{method for defining neighbors, either "threshold" for normalized
distances above a certain value or "nn" for nearest neighbor by normalized distances.}

\item{threshold}{If nearest.method = "threshold" -the value for distance measures 
above which simplify for adjacency matrix}

\item{near.neighbor}{If nearest.method = "nn" - the number of nearest neighbors to use 
in generating an adjacency matrix. If the number of copies of a clone is greater than the
near.neighbor value used, the near.neighbor will randomly return neighbors/clones.}

\item{add.INKT}{Add a additional layer for invariant natural killer T cells based on genes}

\item{add.MAIT}{Add a additional layer for Mucosal-associated invariant T cells based on genes}

\item{n.dim}{The number of Trex dimensions to return, similar to PCA dimensions}

\item{species}{Indicate "human" or "mouse" for gene-based metrics}
}
\value{
Trex eigen vectors caculated from multiplex network
}
\description{
Use this to run the Trex algorithm to return latent vectors
}
