% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.data.R
\name{out.data}
\alias{out.data}
\title{FILL - saves output data}
\usage{
out.data(input, vpd.input, sr.input, prec.input, peak.hours = c(10:14),
  low.sr = 0.15, peak.sr = 0.75, vpd.cutoff = 0.5, prec.lim = 1,
  method = "env.filt", make.plot = TRUE)
}
\arguments{
\item{input}{An \code{\link{is.trex}}-compliant time series from \code{cal.sfd} outputs
(e.g., \code{X$sfd.mw$sfd})}

\item{vpd.input}{An \code{\link{is.trex}}-compliant object (\code{zoo} time-series or \code{data.frame})
with a timestamp and a value column containing the vapour pressure deficit (\emph{vpd}; in kPa)
with the same temporal extent and time_steps as the input data.}

\item{sr.input}{An \code{\link{is.trex}}-compliant object (\code{zoo} time-series or \code{data.frame})
with a timestamp and a value column the solar radiation data (\emph{sr}; e.g., global radiation or \emph{PAR})}

\item{prec.input}{An \code{\link{is.trex}}-compliant object (\code{zoo} time-series or \code{data.frame})
with a timestamp and a value column the precipitation data (\emph{sr}; e.g., global radiation or \emph{PAR})}

\item{peak.hours}{FILL}

\item{low.sr}{FILL}

\item{peak.sr}{FILL}

\item{vpd.cutoff}{FILL}

\item{prec.lim}{FILL}

\item{method}{FILL}

\item{make.plot}{Logical; if \code{TRUE}, a plot is generated presenting the FILL.}
}
\value{
FILL returns output data
}
\description{
FILL this function allow saving plots and making cool data.
}
\examples{
\dontrun{
#Gc response function
raw   <-
  is.trex(
    example.data(type = "doy", species = "PCAB"),
    tz = "GMT",
    time.format = "\%H:\%M",
    solar.time = TRUE,
    long.deg = 7.7459,
    ref.add = FALSE
  )
input <-
  time_step(
    input = raw,
    start = "2013-05-01 00:00",
    end = "2013-11-01 00:00",
    time.int = 15,
    max.gap = 60,
    decimals = 10,
    df = FALSE
  )
input[which(input < 0.2)] <- NA
input <-
  dt.max(
    input,
    methods = c("dr"),
    det.pd = TRUE,
    interpolate = FALSE,
    max.days = 10,
    df = FALSE
  )
output.data <-
  cal.sfd(input,
          make.plot = TRUE,
          df = FALSE,
          wood = "Coniferous")


vpd_raw   <-
  is.trex(
    vpd,
    tz = "GMT",
    time.format = "(\%m/\%d/\%y \%H:\%M:\%S)",
    solar.time = TRUE,
    long.deg = 7.7459,
    ref.add = FALSE
  )
vpd.input <-
  time_step(
    input = vpd_raw,
    start = "2013-05-01 00:00",
    end = "2013-11-01 00:00",
    time.int = 15,
    max.gap = 60,
    decimals = 10,
    df = FALSE
  )
sr_raw   <-
  is.trex(
    sr,
    tz = "GMT",
    time.format = "(\%m/\%d/\%y \%H:\%M:\%S)",
    solar.time = TRUE,
    long.deg = 7.7459,
    ref.add = FALSE
  )
sr.input <-
  time_step(
    input = sr_raw,
    start = "2013-05-01 00:00",
    end = "2013-11-01 00:00",
    time.int = 15,
    max.gap = 60,
    decimals = 10,
    df = FALSE
  )

input <- (output.data$sfd.dr$sfd)


prec.in <- is.trex(preci, time.format = "\%d/\%m/\%y", tz = "UTC")
prec.input <- window(
  prec.in,
  start = as.POSIXct(as.character(zoo::index(input)[1]), format =
                       "\%Y-\%m-\%d", tz = "UTC"),
  end = as.POSIXct(as.character(zoo::index(input)[length(input)]), format =
                     "\%Y-\%m-\%d", tz = "UTC")
)

output <-
  out.data(
    input,
    vpd.input = vpd.input,
    sr.input = sr.input,
    prec.input = prec.input,
    prec.lim = 1,
    method = "env.filt",
    make.plot = TRUE
  )
head(output)
}

}
