% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{outlier}
\alias{outlier}
\title{Data cleaning and outlier detection}
\usage{
outlier()
}
\value{
Once the application is launched,
 the user can load an \code{.RData} file where a \code{data.frame}
  with a imestamp and sensor data (multiple sensor columns are supportred).
  The timestamp in this \code{data.frame} should be of class \code{POSIXct}.
  Users can select the x and y axes of the interactive time series plots.
  In addition, the user can provide the units of the imported data
  (e.g., degrees \eqn{C} or \eqn{mV} for \eqn{\Delta T} or \eqn{\Delta V}, respectively).
  A parameter (alpha) for automatic outlier detection can be supplied.
  More specifically, the automatic identification of outliers is based on a
  two-step procedure:
    i) the Tukey’s method (Tukey, 1977) is applied to detect statistical outliers
    as values falling outside the range
    \eqn{[q_{0.25} - alpha * IQR, q_{0.75} + alpha * IQR]}{[q0.25 - alpha * IQR, q0.75 + alpha * IQR]},
    where \eqn{IQR} is the interquartile range
    (\eqn{q_{0.75} - q_{0.25}}{q0.75 - q0.25})
    with \eqn{q_{0.25}}{q0.25} denoting the 25\% lower quartile and \eqn{q_{0.75}}{q0.25} the
    75\% upper quartile, and alpha is a user-defined parameter
    (default value \code{alpha = 3};
    although visual inspection through the interactive plots allows for adjusting
    alpha and optimizing the automatic detection of outliers),
    and ii) the first difference (or lag-1 differences) of the raw data are calculated
    and data points with lag-1 differences greater
    than the mean of the raw input time series, are excluded.

The function does not return a value, but allows the user
 to save a list containing the raw and outlier-free data, as well as
 the automatically and manually selected outliers in seperate items.
 This object is exported as a ".Rda" or ".Rdata" file and can be
 accessed subsequently via \code{load}.
}
\description{
This function launches a Shiny application that
 (1) visualizes raw and outlier-free time series interactively
 (using {plotly}),
 (2) highlights automatically detected outliers,
 (3) allows the user to revise the automatically detected outliers
    and manually include data points, and
 (4) exports the original data and the outlier-free time series in
    an \code{\link{is.trex}}-compliant object that can be further processed.
}
\examples{
\dontrun{
# find example file path
system.file("exdata", "example.RData", package = "TREX", mustWork = TRUE)
# either copy-paste this into the navigation bar of the file selection window
# or navigate here manually for selection

# launch shiny application
outlier()

}

}
