% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTrex.R
\name{runTrex}
\alias{runTrex}
\title{Trex single cell calculation}
\usage{
runTrex(sc, chains = "TRB", AA.properties = "AF", reduction.name = "Trex")
}
\arguments{
\item{sc}{Single Cell Object in Seurat or SingleCell Experiment format}

\item{chains}{TRA or TRB}

\item{AA.properties}{Amino acid properties to use for distance calculation: 
"AF" = Atchley factors, "KF" = Kidera factors, or "both"}

\item{reduction.name}{Keyword to save Trex reduction. Useful if you want
to try Trex with multiple parameters}
}
\value{
Seurat or SingleCellExperiment object with Trex dimensions placed 
into the dimensional reduction slot.
}
\description{
Run Trex algorithm with Seurat or SingleCellExperiment pipelines
}
\examples{
trex_example <- runTrex(trex_example, 
                        AA.properties = "AF", 
                        reduction.name = "Trex.AF")
                        
}
