% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-TREX.R
\docType{package}
\name{TREX-package}
\alias{TREX}
\alias{TREX-package}
\title{TREX: TRree sap flow EXtractor.}
\description{
Performs data assimilation, processing and analyses on sap flow data obtained with the thermal dissipation method (TDM).
The package includes functions for gap filling time-series data, detecting outliers, calculating data-processing uncertainties and,
generating uniform data output and visualisation. The package is designed to deal with large quantities of data and apply commonly
used data-processing methods. The functions have been validated on data collected from different tree species across the northern hemisphere
(Peters et al. 2018 <doi: 10.1111/nph.15241>).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://the-hull.github.io/TREX/}
  \item Report bugs at \url{https://github.com/the-Hull/TREX/issues}
}

}
\author{
\strong{Maintainer}: Richard Peters \email{richardlouispeters3@hotmail.com} (\href{https://orcid.org/0000-0002-7441-1297}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Christoforos Pappas (\href{https://orcid.org/0000-0001-5721-557X}{ORCID}) [contributor]
  \item Alexander Hurley (\href{https://orcid.org/0000-0002-9641-2805}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Rafael Poyatos [contributor]
  \item Victor Flo [contributor]
  \item Roman Zweifel [contributor]
  \item Willem Goossens [contributor]
  \item Kathy Steppe [thesis advisor, etcher, funder]
}

}
