% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-TREX.R
\docType{package}
\name{TREX}
\alias{TREX}
\alias{TREX-package}
\title{TREX: TRree sap flow EXtractor.}
\description{
Performs data assimilation, processing and analyses on sap flow data obtained with the thermal dissipation method (TDM).
The package includes functions for gap filling time-series data, detecting outliers, calculating data-processing uncertainties,
generating uniform data output and visualisation.
The package is designed to deal with large quantities of data and apply common data-processing methods.
The functions have been validated on data collected from different tree species across the northern hemisphere (Peters et al. 2018 <doi: 10.1111/nph.15241>).
}
